no comment

+++ pyqcstrc/dode/intsct12.pyx
@@ -13,7 +13,10 @@ from pyqcstrc.dode.utils12 cimport triangle_area_6d, obj_area_6d, remove_doublin
 from pyqcstrc.dode.numericalc12 cimport inside_outside_triangle, check_intersection_two_triangles, check_intersection_line_segment_triangle, check_intersection_two_segment_numerical
 
 DTYPE_double = np.float64
-DTYPE_int = int
+#DTYPE_int = int
+DTYPE_int = np.int64
+#ctypedef np.int64_t DTYPE_int_t
+#ctypedef np.float64_t DTYPE_double_t
 
 cdef np.float64_t SIN=np.sqrt(3)/2.0
