+++ pyqcstrc/dode/numericalc12.pyx
@@ -8,7 +8,10 @@ cimport numpy as np
 cimport cython
 
 DTYPE_double = np.float64
-DTYPE_int = int
+#DTYPE_int = int
+DTYPE_int = np.int64
+#ctypedef np.int64_t DTYPE_int_t
+#ctypedef np.float64_t DTYPE_double_t
 
 cdef np.float64_t SIN=np.sqrt(3)/2.0
 cdef np.float64_t EPS=1e-6 # tolerance
